+++ package.json
@@ -6,10 +6,10 @@
   "scripts": {
     "lint": "eslint && prettier --check .",
     "fix": "eslint --fix && prettier --write .",
-    "start": "bun run ./src/index.ts",
-    "start:debug": "DEBUG='*' bun run ./src/index.ts",
-    "serve": "bun run ./src/index.ts serve",
-    "mcp": "bun run ./src/index.ts mcp"
+    "start": "bun run ./server/index.ts",
+    "start:debug": "DEBUG='*' bun run ./server/index.ts",
+    "serve": "bun run ./server/index.ts serve",
+    "mcp": "bun run ./server/index.ts mcp"
   },
   "devDependencies": {
     "@eslint/js": "^9.22.0",
