+++ server/ha-ws-api.test.ts
