+++ server/index.ts
@@ -6,7 +6,7 @@ import { connectToHAWebsocket } from './ha-ws-api'
 import { createBuiltinServers } from './execute-prompt-with-tools'
 import { createDefaultLLMProvider } from './llm'
 import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js'
-import { createHomeAssistantServer } from './servers/home-assistant'
+import { createHomeAssistantServer } from './mcp/home-assistant'
 
 configDotenv()
