+++ server/scene-lerp.test.ts
