+++ shared/prompt.ts
